+++ cmd/stanza/root.go
@@ -17,7 +17,6 @@ import (
 	agent "github.com/observiq/stanza/agent"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 // RootFlags are the root level flags that be provided when invoking stanza from the command line
@@ -31,8 +30,12 @@ type RootFlags struct {
 	MemProfile         string
 	MemProfileDelay    time.Duration
 
-	LogFile string
-	Debug   bool
+	LogLevel      string
+	LogFile       string
+	MaxLogSize    int
+	MaxLogBackups int
+	MaxLogAge     int
+	Debug         bool
 }
 
 // NewRootCmd will return a root level command
@@ -48,11 +51,16 @@ func NewRootCmd() *cobra.Command {
 	}
 
 	rootFlagSet := root.PersistentFlags()
-	rootFlagSet.StringVar(&rootFlags.LogFile, "log_file", "", "write logs to configured path rather than stderr")
+	rootFlagSet.StringVar(&rootFlags.LogLevel, "log_level", "INFO", "sets the agent's log level")
+	rootFlagSet.StringVar(&rootFlags.LogFile, "log_file", "", "writes agent logs to a specified file")
+	rootFlagSet.IntVar(&rootFlags.MaxLogSize, "max_log_size", 10, "sets the maximum size of agent log files in MB before rotating")
+	rootFlagSet.IntVar(&rootFlags.MaxLogBackups, "max_log_backups", 5, "sets the maximum number of rotated log files to retain")
+	rootFlagSet.IntVar(&rootFlags.MaxLogAge, "max_log_age", 7, "sets the maximum number of days to retain a rotated log file")
+	rootFlagSet.BoolVar(&rootFlags.Debug, "debug", false, "debug logging flag - deprecated")
+
 	rootFlagSet.StringSliceVarP(&rootFlags.ConfigFiles, "config", "c", []string{defaultConfig()}, "path to a config file")
 	rootFlagSet.StringVar(&rootFlags.PluginDir, "plugin_dir", defaultPluginDir(), "path to the plugin directory")
 	rootFlagSet.StringVar(&rootFlags.DatabaseFile, "database", "", "path to the stanza offset database")
-	rootFlagSet.BoolVar(&rootFlags.Debug, "debug", false, "debug logging")
 
 	// Profiling flags
 	rootFlagSet.IntVar(&rootFlags.PprofPort, "pprof_port", 0, "listen port for pprof profiling")
@@ -79,12 +87,7 @@ func NewRootCmd() *cobra.Command {
 }
 
 func runRoot(command *cobra.Command, _ []string, flags *RootFlags) {
-	var logger *zap.SugaredLogger
-	if flags.Debug {
-		logger = newDefaultLoggerAt(zapcore.DebugLevel, flags.LogFile)
-	} else {
-		logger = newDefaultLoggerAt(zapcore.InfoLevel, flags.LogFile)
-	}
+	logger := newLogger(*flags).Sugar()
 	defer func() {
 		_ = logger.Sync()
 	}()